cbfstool: Fix tolower() calls on Cygwin
Cygwin complains:

cbfstool.c: 1075:5 error: array subscript has type 'char' [-Werror=char-subscripts]

so add an explicit cast.

Reviewed-on: http://review.coreboot.org/11666
Tested-by: build bot (Jenkins)

+++ util/cbfstool/cbfstool.c
@@ -1219,10 +1219,10 @@ int main(int argc, char **argv)
 				break;
 			case 's':
 				param.size = strtoul(optarg, &suffix, 0);
-				if (tolower(suffix[0])=='k') {
+				if (tolower((int)suffix[0])=='k') {
 					param.size *= 1024;
 				}
-				if (tolower(suffix[0])=='m') {
+				if (tolower((int)suffix[0])=='m') {
 					param.size *= 1024 * 1024;
 				}
 				break;
